backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pdfplumber>=0.9.0
anthropic>=0.7.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
reportlab
